server/pyproject.toml
[project]
name = "server"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appwrite>=15.0.0",
    "flask>=3.1.2",
    "gunicorn>=25.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
